fix a lot of javadoc

+++ ui/src/main/java/alice/util/jedit/TextAreaPainter.java
@@ -98,7 +98,7 @@ public class TextAreaPainter extends JComponent implements TabExpander {
     /**
      * Returns the syntax styles used to paint colorized text. Entry <i>n</i> will be used to paint tokens with id = <i>n</i>.
      *
-     * @see org.gjt.sp.jedit.syntax.Token
+     * @see Token
      */
     public final SyntaxStyle[] getStyles() {
         return styles;
@@ -108,7 +108,7 @@ public class TextAreaPainter extends JComponent implements TabExpander {
      * Sets the syntax styles used to paint colorized text. Entry <i>n</i> will be used to paint tokens with id = <i>n</i>.
      *
      * @param styles The syntax styles
-     * @see org.gjt.sp.jedit.syntax.Token
+     * @see Token
      */
     public final void setStyles(SyntaxStyle[] styles) {
         this.styles = styles;
@@ -338,12 +338,10 @@ public class TextAreaPainter extends JComponent implements TabExpander {
     /**
      * Repaints the text.
      *
-     * @param g The graphics context
+     * @param gfx The graphics context
      */
     public void paint(Graphics gfx) {
-        tabSize = fm.charWidth(' ') * ((Integer) textArea
-                .getDocument().getProperty(
-                        PlainDocument.tabSizeAttribute)).intValue();
+        tabSize = fm.charWidth(' ') * (Integer) textArea.getDocument().getProperty(PlainDocument.tabSizeAttribute);
 
         Rectangle clipRect = gfx.getClipBounds();
